use config file of gscloud with lowest priority
if `uuid` and `token` cannot be found in env vars or in tf definition files, the gscloud config file is used to import `uuid` and `token`. The path of the gscloud config file can be set, or the default path will be used. Only the first configured account in the config file is used.