Fix network ordering & allow assigning DHCP IP to a server (#190)
* deprecate field ordering & change net ordering logic

the network ordering of a server is now based on the defined order of network in the server resource block.

* make ordering field computed

* change shutdown-requiring update logic of server

Only shutdown the server when updating cores, mem, ip addr relations, and/or adding/removing networks to/from the server.

* change server-net relation updating logic

Relink all networks if there are networks being added/removed to/from the server. Otherwise, only update the server-net relations' properties.

* change the server update sequence

UpdateStoragesRel does not need the server to be shutdown. UpdateNetworksRel requires server to be off when the network ordering is changed, or the server-network list is changed. UpdateNetworksRel does not require the server to be off when there are only changes in server-network relations' properties.

* fix the fw rules are not updated

* make network ordering changes require server off

* upgrade gsclient-go v3. 8.0

go mod tidy & go mod vendor

* allow assigning DHCP IP to a server in a network

* reword deprecated message of field ordering

* update docs of server rs

* split the UpdateNetworksRel into 2 methods

make the code easier to read

* make dhcp_range computed

Solve the issue causing terraform to force set dhcp_range to null when dhcp_active=true and dhcp_range is not set. If it is not set and DHCP is enabled, dhcp_range will be set by the backend automatically.

* update network docs