Fix cannot retry upgrade 1.29 to 1.30 when previous upgrade fails (#406)
* ensure freshness of k8s resource after creation

* always get template uuid from version/release even if version/release is not changed

IF the previous upgrade (from 1.29 to 1.30) fails, the updated value of version/release is still persisted (tf provider behaviour because we don't set version/release in k8s resource READ). Hence when we re-try the upgrade, this piece of code detect no changes in version/release parameter => tf tries to get template uuid from service_template_uuid (which is just the current template of the cluster), hence no upgrade happens.